minor bug fix

+++ Hosinsa/src/main/java/com/hosinsa/controller/AdminController.java
@@ -228,7 +228,7 @@ public class AdminController {
 	}
 	
 	@GetMapping("/order/{orderNum}")
-	public String getOrderDetail(@PathVariable int orderNum,Model model) {
+	public String getOrderDetail(@PathVariable long orderNum,Model model) {
 		model.addAttribute("order",adminService.getOrder(orderNum));
 		return "/admin/order";
 	}
