+++ Hosinsa/src/main/java/com/hosinsa/service/AdminService.java
@@ -19,6 +19,6 @@ public interface AdminService {
 	public List<MemberVO> searchMember(Criteria cri);
 	public int searchTotal(Criteria cri);
 	public List<OrderVO> getOrderList(String process);
-	public OrderVO getOrder(int orderNum);
+	public OrderVO getOrder(long orderNum);
 	public boolean updateProcess(OrderVO vo);
 }
